fix: store typings in separate file
